Update2: Remove the content related to SUBGRID_CD_FIXED from the CALCULATE_SUBGRID_CD subroutine
This seems a logical bug when computing Cdu and Cdv but it's not. Defautly, Cdu and Cdv has been
assigned as Cd_fixed (SUBGRID_CD_FIXED is default option).
There are in total 4 methods to compute the bottom Cd:
1. SUBGRID_CD_FIXED: if defined, a Cd_fixed is required (default=0.002)
2. MANNING_TYPE = 'UNIFORM': if defined, a constant Manning is required (default=0.01)
3. MANNING_TYPE = 'VARY': if defined, a file for ManningVary matrix is required.
4. MANNING_TYPE = 'SUBGRID': if defined, a file for ManningSubGrid is required.
In the io.F, only one of these 4 methods can be TRUE at a time. Therefore, it would be much better
if each option is isolated.